refactor(custom-cmds): address /code-review findings — dedup cache IO, share types, fix color map
